+++ core/src/main/java/org/apache/mahout/cf/taste/hadoop/SlopeOnePrefsToDiffsJob.java
@@ -20,7 +20,7 @@ package org.apache.mahout.cf.taste.hadoop;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.InputFormat;
 import org.apache.hadoop.mapreduce.Job;
@@ -69,7 +69,7 @@ public final class SlopeOnePrefsToDiffsJob extends Job {
 
     jobConf.setClass("mapred.reducer.class", SlopeOnePrefsToDiffsReducer.class, Reducer.class);
     jobConf.setClass("mapred.output.key.class", ItemItemWritable.class, Object.class);
-    jobConf.setClass("mapred.output.value.class", DoubleWritable.class, Object.class);
+    jobConf.setClass("mapred.output.value.class", FloatWritable.class, Object.class);
 
     jobConf.setClass("mapred.output.format.class", SequenceFileOutputFormat.class, OutputFormat.class);
     jobConf.set("mapred.output.dir", StringUtils.escapeString(outputPathPath.toString()));
